Fix debug build without release keystore or password.
All signing configuration blocks are always evaluated regardless of
which config is going to be used. This means, that even for debug usage,
password is resolved by running security unless running build in Travis.
This causes an extra problem since 'security' is a OSX-specific command.
Also if storeFile is set, it must be an existing file.

- Ignore password resolution exception to allow building without OSX.
- Refrain from setting undefined storeFile.